new wall variant, windows as layers, rearrange
Revoked, brick and wooden walls are now rearranged and won't
contain a window variant. The windows are now separated, so
they can be reutilized on a separated layer, no matter the
wall type.
The width of the tileset has been incrased from 8 to 12 tiles:
at the left are the walls and windows, at the right the floors
with an empty space for variants.